Refactor: create src folder and move files
